requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pytube==15.0.0
python-multipart==0.0.6
email-validator==2.0.0
ffmpeg-python==0.2.0